reddit client
This is a jabbascript client for reddit chat.  It works!  But it's
not useful without a controlling Node.JS application that inserts
it into a reddit.com/chat/r page.  That application also needs to
manually handle logging in if the browser is redirected to /login

I have however actually tested this code unlike some of my unfinished
commits and it does in fact work for both sending and receiving
messages.